fix: remove circular dependencies in multiple files
