+++ packages/react-vapor/src/components/checkbox/GroupableCheckboxConstants.ts
@@ -1,0 +1,12 @@
+import {ICheckboxState} from './CheckboxReducers';
+
+export interface IGroupableCheckboxesState {
+    total: number;
+    nbChecked: number;
+    parentId?: string;
+    parent?: ICheckboxState;
+    checkboxes: ICheckboxState[];
+}
+
+export const groupableCheckboxInitialState: IGroupableCheckboxesState = {total: 0, nbChecked: 0, checkboxes: []};
+export const groupableCheckboxesInitialState: IGroupableCheckboxesState[] = [];
