+++ packages/react-vapor/src/components/datePicker/DatePickerConstants.ts
@@ -4,3 +4,13 @@
  * and Array[1] represents the upper limit Date
  */
 export type DatePickerDateRange = [Date, Date];
+
+export const DatePickerColors = {
+    blue: 'blue',
+    green: 'green',
+    yellow: 'yellow',
+    red: 'red',
+    orange: 'orange',
+};
+
+export const DEFAULT_DATE_PICKER_COLOR: string = DatePickerColors.blue;
