+++ packages/react-vapor/src/components/dropdownSearch/MultiSelectDropdownSearch/index.ts
@@ -1,6 +1,5 @@
 export * from './DraggableSelectedOption';
 export * from './MultiSelectDropdownSearch';
 export * from './MultiSelectDropdownSearchConnected';
-export * from './MultiSelectDropdownSearchReducer';
 export * from './MultiSelectInput';
 export * from './SelectedOption';
