+++ packages/react-vapor/src/components/pagination/PaginationConstants.ts
@@ -1,0 +1,4 @@
+export const TablePaginationDefaultValue = {
+    PerPage: [10, 20, 30],
+    perPageLabel: 'Results per page',
+};
