+++ packages/react-vapor/src/components/pagination/TablePagination.tsx
@@ -38,8 +38,3 @@ export const TablePagination: FunctionComponent<ITablePaginationProps> = ({
         />
     </div>
 );
-
-export const TablePaginationDefaultValue = {
-    PerPage: [10, 20, 30],
-    perPageLabel: 'Results per page',
-};
