+++ packages/react-vapor/src/components/select/SelectConstants.ts
@@ -1,0 +1,3 @@
+export const SelectConstants = {
+    DropGroupId: 'select',
+};
