+++ packages/react-vapor/src/components/select/index.ts
@@ -1,3 +1,4 @@
+export * from './SelectConstants';
 export * from './MultiSelectConnected';
 export * from './SelectActions';
 export * from './SelectConnected';
