+++ packages/react-vapor/src/components/tab/TabConstants.ts
@@ -1,0 +1,6 @@
+export const TabConstants = {
+    DefaultGroupId: 'default',
+};
+
+/* @deprecated use TabConstants.DefaultGroupId instead */
+export const DEFAULT_GROUP_ID = TabConstants.DefaultGroupId;
