+++ packages/react-vapor/src/components/tab/TabSelectors.ts
@@ -1,12 +1,12 @@
 import {createSelector} from 'reselect';
 import _ from 'underscore';
 
-import {DEFAULT_GROUP_ID, ITabPaneOwnProps} from '.';
 import {IReactVaporState} from '../../ReactVapor';
 import {ITabGroupState} from './TabReducers';
+import {TabConstants} from './TabConstants';
 
-const getTabGroup = (state: IReactVaporState, ownProps: ITabPaneOwnProps) =>
-    _.findWhere(state.tabs, {id: ownProps.groupId ?? DEFAULT_GROUP_ID});
+const getTabGroup = (state: IReactVaporState, {groupId}: {groupId?: string}) =>
+    _.findWhere(state.tabs, {id: groupId ?? TabConstants.DefaultGroupId});
 
 const getSelectedTab = createSelector(getTabGroup, (tabGroup: ITabGroupState) =>
     _.findWhere(tabGroup?.tabs, {isSelected: true})
